Fix keyboard selection jitter during hover
Previously, if you used arrows or home/end to navigate a menu/folder
that was taller than the available space, and the mouse was over any
part of the menu, the selection would jump to the item at the mouse
position (because the browser issues a new mouseover event when the
items scroll up or down).

This release fixes that issue by ignoring mouseover events when the
mouse hasn't moved relative to the screen.